requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
idna==2.8
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.2
django-cors-headers==3.0.1
psycopg2
django-proxy
matplotlib==3.1.0
rasterio==1.0.24
Shapely==1.6.4.post2
torch==1.1.0
pyshp==2.1.0
scipy==1.3.0